# SWAGGER_API_DOCUMENTATION.yaml
openapi: 3.0.0
info:
  title: Lilium B2B E-commerce Backend API
  description: Complete API documentation for the Lilium B2B multi-vendor platform with cash-on-delivery system
  version: 1.0.0
  contact:
    name: Lilium API Support
    email: support@lilium.iq
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.lilium.iq/api
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Promotion:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        type:
          type: string
          enum: [PERCENTAGE, FIXED_AMOUNT]
        value:
          type: number
        maxDiscount:
          type: number
        minPurchase:
          type: number
        targetType:
          type: string
          enum: [ALL, PRODUCT, CATEGORY]
        zones:
          type: array
          items:
            type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        isActive:
          type: boolean
paths:
...
  /promotions:
    get:
      tags: [Promotions]
      summary: List promotions
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
        - name: zone
          in: query
          schema:
            type: string
      responses:
        200:
          description: List of promotions

    post:
      tags: [Promotions]
      summary: Create promotion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Promotion'
      responses:
        201:
          description: Promotion created
  /promotions/{id}:
    get:
      tags: [Promotions]
      summary: Get promotion details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Promotion details

    put:
      tags: [Promotions]
      summary: Update promotion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Promotion'
      responses:
        200:
          description: Promotion updated

    delete:
      tags: [Promotions]
      summary: Delete promotion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Promotion deleted
...